Move FlightsList into details folder
