+++ client/src/components/details/DetailsTab.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import { fetchDestinationCategory } from '../../actions';
 import AppBar from 'material-ui/AppBar';
 import Tabs, { Tab } from 'material-ui/Tabs';
-import FlightsList from '../FlightsList';
+import FlightsList from './FlightsList';
 
 function TabContainer(props) {
   return (
